git-svn-id: https://svn.apache.org/repos/asf/hive/branches/spark@1623263 13f79535-47bb-0310-9956-ffa450edef68

+++ service/src/java/org/apache/hive/service/cli/thrift/ThriftCLIService.java
@@ -61,7 +61,7 @@ public abstract class ThriftCLIService extends AbstractService implements TCLISe
 
   protected int minWorkerThreads;
   protected int maxWorkerThreads;
-  protected int workerKeepAliveTime;
+  protected long workerKeepAliveTime;
 
   protected static HiveAuthFactory hiveAuthFactory;
